FEAT: Se implementa navegacion por pestañas en la vista indiviual del producto
