+++ src/pages/productos/[id].astro
@@ -1,13 +1,11 @@
 ---
 import { GetProductId } from "@features/productos/services";
 import SidebarLayout from "@shared/layouts/SidebarLayout.astro";
-import { ProductCard } from "@features/productos/components/ProductCard/ProductCard";
 import { AnalisisSelect } from "@features/productos/components/AnalisisSelect/AnalisisSelect";
-import { EntryDetailsCard } from "@features/productos/components/EntryDetailsCard/EntryDetailsCard";
-import { AnalisisOutlet } from "@features/productos/components/AnalisisOutlet/AnalisisOutlet";
+
 import PillHeading from "@shared/components/PillHeading/PillHeading.astro";
 
-import { TabsContainer, Tab } from "@shared/components";
+import { TabsContainer } from "@features/productos/components/TabsContainer/TabsContainer.jsx";
 // Se hace trae la informacion del producto desde la API
 const { id } = Astro.params;
 let data = null;
@@ -23,47 +21,20 @@ try {
 <SidebarLayout>
     <div class="container mainContainer">
 
-        <TabsContainer data={data} client:load/>
-
-        <!-- Titulo
         <div class="header">
             <PillHeading variant="primary">
                 <h1>Informacion del Producto</h1>
-            </PillHeading>
-            {data && (
-                <AnalisisSelect
+                </PillHeading>
+                {data && (
+                    <AnalisisSelect
                     numeroRegistroProducto={data.numeroRegistroProducto}
                     client:load
-                />
-            )}
+                    />
+                )}
         </div>
 
-        <section id="productInfo">
-            <div class="row">
-                <div class="col-lg-6">
-                    {
-                        data && (
-                            <ProductCard client:load producto={data.producto} />
-                        )
-                    }
-                </div>
-                <div class="col-lg-6">
-                    {data && <EntryDetailsCard client:load data={data} />}
-                </div>
-            </div>
-        </section>
-
-        <section class="analisisSection">
-            <h2>Analisis Registrados</h2>
-
-            {data && (
-                <AnalisisOutlet
-                    numeroRegistroProducto={data.numeroRegistroProducto}
-                    client:visible
-                />
-            )}
-        </section> -->
-
+        <TabsContainer data={data} client:load/>
+        
     </div>
 </SidebarLayout>
 <style>
