+++ src/shared/components/Tabs/Tab/Tab.module.css
@@ -1,11 +1,17 @@
+
 .tab {
-    padding: 12px 16px;
+    padding: 7px 14px;
     cursor: pointer;
-    border-bottom: 2px solid transparent;
-    transition: border-color 0.3s;
+    border: 1px solid transparent;
+    border-radius: 10px 10px 0px 0px;
+    transition: 0.3s all;
+    background: #ececec;
+    border: 1px solid #e7e6e6;
 }
 
 .tab.active {
-    border-color: var(--color-primario);
+    border: 1px solid #e7e6e6;
+    border-bottom: none;
     font-weight: bold;
+    background: white;
 }
