--- src/shared/components/Tabs/TabContent/TabContent.jsx
@@ -1,9 +1,0 @@
-import styles from './TabContent.module.css';
-
-export const TabContent = ({ children, isActive=false }) => {
-  return (
-    <div className={`${styles.tabContent} ${isActive ? styles.active : ''}`}>
-      {children}
-    </div>
-  );
-}
