--- src/shared/components/Tabs/TabContent/TabContent.module.css
@@ -1,11 +1,0 @@
-.tabContent {
-    padding: 20px;
-    border: 1px solid var(--color-border);
-    border-radius: 5px;
-    background-color: var(--color-background);
-    display: none;
-}
-
-.active {
-    display: block;
-}
