--- src/shared/components/Tabs/TabsContainer/TabsContainer.module.css
@@ -1,22 +1,0 @@
-.tabsContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  background-color: white;
-}
-
-.tabsList {
-  display: flex;
-  gap: 1rem;
-  padding: 0 1rem;
-  border-bottom: 2px solid var(--color-primario-claro);
-}
-
-.tabWrapper {
-  cursor: pointer;
-}
-
-.contentContainer {
-  padding: 1rem;
-}
