+++ src/shared/components/index.js
@@ -17,5 +17,4 @@ export * from "./AnalisisCard/AnalisisCard.jsx"
 export * from "./Loaders/LoaderSpiner/LoaderSpiner.jsx"
 export * from "./Loaders/ComponentLoader/ComponentLoader.jsx"
 export * from "./inputs/CustomCheckBox/CustomCheckBox.jsx"
-export * from "./Tabs/TabsContainer/TabsContainer.jsx"
 export * from "./Tabs/Tab/Tab.jsx"
